Move yield-fixture set_db to conftest.py and add it in test_post_return_user_workspace.
