+++ tests/test_post_login.py
@@ -2,19 +2,7 @@ import pytest, json
 from unittest import mock
 from flask import current_app
 from main import logger
-from models import User
 
-
-@pytest.fixture(autouse=True)
-def set_db(app):
-    with app.app_context():
-        test_session = current_app.Session()
-        test_user = User(name='denchik', passwordhash='pbkdf2:sha256:150000$wHwsgiLd$6979f267446c0e3d2797c21006f9272c3e19d5c70925d87989983ab3826350d8', useremail='foo@bar.baz')
-        test_session.add(test_user)
-        test_session.commit()
-        yield None
-        test_session.delete(test_user)
-        test_session.commit()
 
 @pytest.mark.parametrize(
         ('data', 'headers', 'logger_message', 'message', 'http_code'),
@@ -31,7 +19,7 @@ def set_db(app):
         ),
         ids = ['No Content-Type','Content-Type is xml','Not json format','No password and data', 'Not string name', 'Not string password', 'Not registered name', 'Invalid password']
 )
-def test_login_with_invalid_data_and_headers(client, app, data, headers, logger_message, message, http_code):
+def test_login_with_invalid_data_and_headers(client, app, set_db, data, headers, logger_message, message, http_code):
         with mock.patch.object(logger, 'info') as mock_info:
                 response = client.post(
                         '/login', data = data, headers = headers)
@@ -40,7 +28,7 @@ def test_login_with_invalid_data_and_headers(client, app, data, headers, logger_
                 assert message in response.data
                 assert http_code == response.status_code
 
-def test_login_with_valid_data_and_header(client, app):
+def test_login_with_valid_data_and_header(client, app, set_db):
         response = client.post(
                         '/login', data = json.dumps({"name": "denchik", "password": "foobar"}), headers = {'content-type': 'application/json'})
